extracted transaction history into a separate ui component
